refactor: separated Avatar & AvatarForm
