+++ expo/components/profile/Avatar.tsx
@@ -1,122 +1,49 @@
-import { decode } from "base64-arraybuffer";
-import * as FileSystem from "expo-file-system";
-import * as ImagePicker from "expo-image-picker";
 import { forwardRef, useEffect, useImperativeHandle, useState } from "react";
-import {
-  Alert,
-  Image,
-  Platform,
-  Pressable,
-  StyleSheet,
-  View,
-} from "react-native";
+import { Image, StyleSheet, View } from "react-native";
 
 import { supabase } from "../../lib/supabase";
-import { useSupabaseUserHook } from "../../lib/useSupabaseUser";
-import Button from "../Button";
-import { formStyles } from "../ControlledInput";
-import { Text } from "../Themed";
 
-interface AvatarProps {
-  onImageLoad: () => void;
-}
+export type AvatarRemote = {
+  refresh: () => void;
+};
+type AvatarProps = {
+  id: string | undefined;
+  tempoAvatarImage?: string; // if we want to pass an image url directly (on edit profile)
+};
 
-/**
- * Component with a button to select an image from the device and display it.
- * It also has a saveImage method to upload the image to Supabase Storage.
- */
-const Avatar = forwardRef((props: AvatarProps, ref) => {
-  const { onImageLoad } = props;
-  const user = useSupabaseUserHook();
-  const [uploading, setUploading] = useState(false);
-  const [avatarUrl, setAvatarUrl] = useState<string | null>(null);
+const Avatar = forwardRef<AvatarRemote, AvatarProps>(
+  ({ id, tempoAvatarImage }, ref) => {
+    const [avatarUrl, setAvatarUrl] = useState<string | null>(null);
 
-  useEffect(() => {
-    if (user) {
+    useImperativeHandle(ref, () => ({
+      refresh: async () => {
+        await downloadUserImage();
+      },
+    }));
+
+    useEffect(() => {
+      if (!id) return;
       downloadUserImage();
-    }
-  }, [user]);
+    }, [id]);
 
-  useImperativeHandle(ref, () => ({
-    saveImage: async () => {
-      return await uploadAvatar();
-    },
-  }));
+    async function downloadUserImage() {
+      const path = `${id}.jpg`;
+      const { data } = supabase.storage
+        .from("avatars")
+        .getPublicUrl(path + "?avoidCache=" + Math.random());
 
-  async function downloadUserImage() {
-    if (!user) return;
-
-    const path = `${user.id}.jpg`;
-    const { data } = supabase.storage
-      .from("avatars")
-      .getPublicUrl(path + "?avoidCache=" + Math.random());
-
-    if (!data) {
-      // No image for this user
-      return;
-    }
-    setAvatarUrl(data.publicUrl);
-  }
-
-  async function uploadAvatar(): Promise<{
-    error: string | null;
-  }> {
-    if (!user || !avatarUrl) {
-      return {
-        error: "No user or avatar",
-      };
-    }
-    const fileName = `${user.id}.jpg`;
-    const image = await getBase64Image(avatarUrl);
-
-    const { error } = await supabase.storage
-      .from("avatars")
-      .upload(fileName, image, {
-        upsert: true,
-        contentType: "image/jpeg",
-      });
-
-    if (error) {
-      console.log("error", error);
-      return {
-        error: error.message,
-      };
-    }
-    downloadUserImage();
-    return {
-      error: null,
-    };
-  }
-
-  async function selectAvatar() {
-    if (!user) return;
-    setUploading(true);
-
-    const file = await ImagePicker.launchImageLibraryAsync({
-      allowsEditing: true,
-      quality: 1,
-      aspect: [1, 1],
-    });
-
-    if (file.assets) {
-      if (file.assets.length !== 1) {
-        Alert.alert("Veuillez sélectionner une seule image");
+      if (!data) {
+        // No image for this user
         return;
       }
-      const image = file.assets[0];
-      setAvatarUrl(image.uri);
-      onImageLoad();
+      setAvatarUrl(data.publicUrl);
     }
-    setUploading(false);
-  }
 
-  return (
-    <View style={{ width: "100%", gap: 10 }}>
-      <Text style={formStyles.label}>Photo de profil</Text>
-      <Pressable onPress={selectAvatar}>
+    return (
+      <>
         {avatarUrl ? (
           <Image
-            source={{ uri: avatarUrl }}
+            source={{ uri: tempoAvatarImage ?? avatarUrl }}
             aria-aria-label="Avatar"
             style={[
               { aspectRatio: 1, width: "100%" },
@@ -133,21 +60,10 @@ const Avatar = forwardRef((props: AvatarProps, ref) => {
             ]}
           />
         )}
-      </Pressable>
-      <Button
-        onPress={selectAvatar}
-        disabled={uploading}
-        block
-        type="outline"
-        size="small"
-      >
-        {uploading ? "Chargement ..." : "Charger une image"}
-      </Button>
-    </View>
-  );
-});
-
-export default Avatar;
+      </>
+    );
+  }
+);
 
 const styles = StyleSheet.create({
   avatar: {
@@ -166,17 +82,4 @@ const styles = StyleSheet.create({
   },
 });
 
-const getBase64Image = async (uri: string) => {
-  if (Platform.OS === "web") {
-    const uriWithoutMIME = uri.split(",")[1];
-    const base64data = decode(uriWithoutMIME);
-    return base64data;
-  } else {
-    // "uri": "file:///data/user/0/host.exp.exponent/cache/DocumentPicker/189be1eb-08a5-4bcb-8001-28aafd0febd6.jpg"}],
-    const fileUri = uri;
-    const base64Data = await FileSystem.readAsStringAsync(fileUri, {
-      encoding: FileSystem.EncodingType.Base64,
-    });
-    return decode(base64Data);
-  }
-};
+export default Avatar;
